+++ db/migrate/20191206142516_add_access_token_to_users.rb
@@ -1,0 +1,5 @@
+class AddAccessTokenToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :token, :string
+  end
+end